feat: Rename compiler plugins, use single context object, retain front matter.
