+++ packages/idyll-compiler/src/index.js
@@ -6,57 +6,50 @@ const parseFrontMatter = require('gray-matter');
 const cleanNewlines = require('./util/clean-newlines');
 const pipeline = require('./util/pipeline');
 
-const astTransformers = [
-  require('./transformers/hoist-variables'),
-  require('./transformers/flatten-children'),
-  require('./transformers/make-full-width'),
-  require('./transformers/wrap-text'),
-  require('./transformers/clean-results'),
-  require('./transformers/smart-quotes'),
-  require('./transformers/auto-linkify')
+const defaultPlugins = [
+  require('./plugins/hoist-variables'),
+  require('./plugins/flatten-children'),
+  require('./plugins/make-full-width'),
+  require('./plugins/wrap-text'),
+  require('./plugins/clean-results'),
+  require('./plugins/smart-quotes'),
+  require('./plugins/auto-linkify')
 ];
 
-module.exports = function(input, options, alias, callback) {
+module.exports = function(input, context = {}) {
   // prepare compiler options
-  options = Object.assign(
-    {},
-    { spellcheck: false, smartquotes: true },
-    options || {}
-  );
+  context = Object.assign({ spellcheck: false, smartquotes: true }, context);
 
   // pre-process input text
   input = cleanNewlines(input).trim();
 
-  // parse YAML front matter, discard for now
+  // parse YAML front matter
   const { content, data } = parseFrontMatter(input);
+  context.metadata = data || {};
 
   // perform lexing
-  const lex = lexer({}, alias);
+  const lex = lexer({}, context.alias);
   let lexResults = '';
   try {
     lexResults = lex(content);
   } catch (err) {
     console.warn(`\nError parsing Idyll markup:\n${err.message}`);
-    return new Promise((resolve, reject) => reject(err));
+    throw err;
   }
 
   // perform parsing to construct an Idyll AST
   let ast;
   try {
-    ast = parse(content, lexResults.tokens, lexResults.positions, options);
+    ast = parse(content, lexResults.tokens, lexResults.positions);
     ast = convertV1ToV2(ast);
   } catch (err) {
     console.warn(`\nError parsing Idyll markup:\n${err.message}`);
-    if (options.async) {
-      return new Promise((resolve, reject) => reject(err));
-    } else {
-      throw err;
-    }
+    throw err;
   }
 
   // construct AST transformation pipeline
-  const transform = pipeline(astTransformers, options.postProcessors || []);
+  const transform = pipeline(defaultPlugins, context.plugins || []);
 
   // apply AST transformations and return result
-  return transform(ast, options);
+  return transform(ast, context);
 };
