+++ packages/idyll-compiler/src/plugins/clean-results.js
