+++ packages/idyll-compiler/src/plugins/flatten-children.js
